# bin/build/swagger/swagger.json
{
	"basePath": "/v1",
	"consumes": [
		"application/json"
	],
	"definitions": {
		"MContentView": {
			"properties": {
				"title": {
					"type": "string"
				},
				"description": {
					"type": "string"
				},
				"content": {
					"type": "string"
				},
				"active": {
					"type": "number",
					"format": "double"
				},
				"image": {
					"type": "string"
				},
				"categories": {
					"type": "string"
				},
				"userId": {
					"type": "string"
				}
			},
			"required": [
				"title",
				"description",
				"content",
				"active",
				"image",
				"categories",
				"userId"
			],
			"type": "object"
		},
		"IPaginationModel": {
			"properties": {
				"count": {
					"type": "number",
					"format": "double"
				},
				"page": {
					"type": "number",
					"format": "double"
				},
				"limit": {
					"type": "number",
					"format": "double"
				},
				"totalPages": {
					"type": "number",
					"format": "double"
				},
				"docs": {
					"type": "array",
					"items": {
						"type": "object"
					}
				}
			},
			"required": [
				"count",
				"page",
				"limit",
				"totalPages",
				"docs"
			],
			"type": "object"
		},
		"MCategoryView": {
			"properties": {
				"index": {
					"type": "number",
					"format": "double"
				},
				"name": {
					"type": "string"
				},
				"description": {
					"type": "string"
				},
				"active": {
					"type": "number",
					"format": "double"
				}
			},
			"required": [
				"index",
				"name",
				"description",
				"active"
			],
			"type": "object"
		},
		"MTagView": {
			"properties": {
				"contentId": {
					"type": "string"
				},
				"categoryId": {
					"type": "string"
				}
			},
			"required": [
				"contentId",
				"categoryId"
			],
			"type": "object"
		},
		"TagEntity": {
			"properties": {
				"_id": {
					"type": "object"
				},
				"created": {
					"type": "number",
					"format": "double",
					"x-nullable": true
				},
				"updated": {
					"type": "number",
					"format": "double",
					"x-nullable": true
				},
				"contentId": {
					"type": "string"
				},
				"categoryId": {
					"type": "string"
				}
			},
			"required": [
				"_id",
				"contentId",
				"categoryId"
			],
			"type": "object"
		}
	},
	"info": {
		"title": "@tn/baithuocnam.service",
		"version": "1.0.0",
		"description": "BaiThuocNamHayService",
		"license": {
			"name": "MIT"
		}
	},
	"paths": {
		"/content/{id}": {
			"get": {
				"operationId": "GetById",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/MContentView"
						}
					}
				},
				"tags": [
					"Content"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string"
					}
				]
			},
			"delete": {
				"operationId": "Delete",
				"produces": [
					"application/json"
				],
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Content"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string"
					}
				]
			}
		},
		"/content": {
			"get": {
				"operationId": "GetPaginated",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/IPaginationModel"
						}
					}
				},
				"tags": [
					"Content"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": true,
						"format": "double",
						"type": "number"
					},
					{
						"in": "query",
						"name": "limit",
						"required": true,
						"format": "double",
						"type": "number"
					},
					{
						"in": "query",
						"name": "fields",
						"required": false,
						"type": "string"
					},
					{
						"in": "query",
						"name": "sort",
						"required": false,
						"type": "string"
					},
					{
						"in": "query",
						"name": "q",
						"required": false,
						"type": "string"
					}
				]
			},
			"post": {
				"operationId": "AddContent",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/MContentView"
						}
					}
				},
				"tags": [
					"Content"
				],
				"security": [],
				"parameters": [
					{
						"in": "body",
						"name": "content",
						"required": true,
						"schema": {
							"$ref": "#/definitions/MContentView"
						}
					}
				]
			}
		},
		"/Category/{id}": {
			"get": {
				"operationId": "GetById",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/MCategoryView"
						}
					}
				},
				"tags": [
					"Category"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string"
					}
				]
			},
			"delete": {
				"operationId": "Delete",
				"produces": [
					"application/json"
				],
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Category"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string"
					}
				]
			}
		},
		"/Category": {
			"get": {
				"operationId": "GetPaginated",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/IPaginationModel"
						}
					}
				},
				"tags": [
					"Category"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": true,
						"format": "double",
						"type": "number"
					},
					{
						"in": "query",
						"name": "limit",
						"required": true,
						"format": "double",
						"type": "number"
					},
					{
						"in": "query",
						"name": "fields",
						"required": false,
						"type": "string"
					},
					{
						"in": "query",
						"name": "sort",
						"required": false,
						"type": "string"
					},
					{
						"in": "query",
						"name": "q",
						"required": false,
						"type": "string"
					}
				]
			},
			"post": {
				"operationId": "AddContent",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/MCategoryView"
						}
					}
				},
				"tags": [
					"Category"
				],
				"security": [],
				"parameters": [
					{
						"in": "body",
						"name": "content",
						"required": true,
						"schema": {
							"$ref": "#/definitions/MCategoryView"
						}
					}
				]
			}
		},
		"/tag/{id}": {
			"get": {
				"operationId": "GetById",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/MTagView"
						}
					}
				},
				"tags": [
					"Tag"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string"
					}
				]
			},
			"delete": {
				"operationId": "Delete",
				"produces": [
					"application/json"
				],
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Tag"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string"
					}
				]
			}
		},
		"/tag": {
			"get": {
				"operationId": "GetPaginated",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/IPaginationModel"
						}
					}
				},
				"tags": [
					"Tag"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": true,
						"format": "double",
						"type": "number"
					},
					{
						"in": "query",
						"name": "limit",
						"required": true,
						"format": "double",
						"type": "number"
					},
					{
						"in": "query",
						"name": "fields",
						"required": false,
						"type": "string"
					},
					{
						"in": "query",
						"name": "sort",
						"required": false,
						"type": "string"
					},
					{
						"in": "query",
						"name": "q",
						"required": false,
						"type": "string"
					}
				]
			},
			"post": {
				"operationId": "AddContent",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/TagEntity"
						}
					}
				},
				"tags": [
					"Tag"
				],
				"security": [],
				"parameters": [
					{
						"in": "body",
						"name": "tag",
						"required": true,
						"schema": {
							"$ref": "#/definitions/MTagView"
						}
					}
				]
			}
		}
	},
	"produces": [
		"application/json"
	],
	"swagger": "2.0",
	"securityDefinitions": {},
	"host": "localhost:3000"
}